requirements.txt
pip~=20.1.1
wheel~=0.34.2
py~=1.8.1
setuptools~=41.2.0
fasttext~=0.9.2
gensim~=3.8.3
transformers~=2.11.0
blingfire~=0.1.3
sentencepiece~=0.1.91
nltk~=3.5
mysql-connector-python~=8.0.21
scipy~=1.4.1
sklearn
pandas
xlrd
psutil
matplotlib
pylint
autopep8